requirements.txt
asgiref==3.4.1
certifi==2022.6.15
charset-normalizer==2.0.12
Django==3.2.14
djangorestframework==3.13.1
idna==3.3
openbabel==2.4.0
Pillow==8.4.0
psycopg2==2.9.3
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
typing-extensions==4.1.1
urllib3==1.26.9
